requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.8.0
python-dotenv>=1.0.0
subprocess32>=3.5.0; python_version < '3.8'
pydantic>=2.5.0
click>=8.1.0